+++ pyfuntofem/interface/__init__.py
@@ -30,5 +30,13 @@ limitations under the License.
 from .driver import *
 from .model import *
 from .optimization import *
-from .interface import *
+from .solver_interface import *
 
+# Import all of the funtofem interfaces
+# underscores are used here to prevent conflicting imports with fun3d, tacs, etc.
+from ._cart3d import *
+from ._fun3d import *
+#from ._openmdao import *
+from ._su2 import *
+from ._tacs import *
+from ._test import *
