+++ pyfuntofem/interface/_cart3d/cart3d_interface.py
@@ -27,7 +27,7 @@ __all__ = ["Cart3DInterface"]
 import numpy as np
 import os
 from funtofem import TransferScheme
-from ..solver_interface import SolverInterface
+from pyfuntofem.interface.solver_interface import SolverInterface
 from .cart3d_utils import ReadTriangulation, ComputeAeroLoads, WriteTri, RMS
